hurd: Fix tcflag_t and speed_t types on 64-bit
These are supposed to stay 32-bit even on 64-bit systems. This matches
BSD and Linux, as well as how these types are already defined in
tioctl.defs

+++ bits/termios.h
@@ -99,13 +99,13 @@
    `tcflag_t', `cc_t', `speed_t' and the `TC*' constants appropriately.  */
 
 /* Type of terminal control flag masks.  */
-typedef unsigned long int tcflag_t;
+typedef unsigned int tcflag_t;
 
 /* Type of control characters.  */
 typedef unsigned char cc_t;
 
 /* Type of baud rate specifiers.  */
-typedef long int speed_t;
+typedef int speed_t;
 
 /* Terminal control structure.  */
 struct termios
